refactor(api-server): extract inspect handler
Move the resource inspect endpoints (matching dataplanes, proxy
config, rules, and the policy-config handlers) out of the
resourceEndpoints god-struct into a dedicated resourceInspectHandler in
resource_inspect_endpoints.go.

Both types embed the shared resourceEndpointsContext, so the moved
method bodies are unchanged except for the receiver. resourceEndpoints
keeps route registration and delegates the inspect routes to its
inspect handler; the inspect-only dependencies (meshContextBuilder,
xdsHooks, knownInternalAddresses) move with it.

resource_endpoints.go drops from ~1500 to ~780 lines. Behavior-
preserving (structural only); the full api-server suite passes.
